Action on github v.02
JIRA: noissue
